modem: cellular: move individual modems to dedicated files
Simplify the `modem_cellular.c` file by pulling out all the individual
modem macros and instantiation to a dedicated file per modem.
