+++ include/zephyr/drivers/modem/modem_cellular.h
@@ -318,6 +318,43 @@ void modem_cellular_chat_on_modem_ready(struct modem_chat *chat, char **argv, ui
 				   (,), inst, __VA_ARGS__)                                         \
 	);
 
+/* Define common chat matches for cellular modems to reduce copy-pasting */
+#define MODEM_CELLULAR_COMMON_CHAT_MATCHES()							   \
+	MODEM_CHAT_MATCH_DEFINE(ok_match, "OK", "", NULL);					   \
+	MODEM_CHAT_MATCHES_DEFINE(__maybe_unused allow_match,					   \
+				  MODEM_CHAT_MATCH("OK", "", NULL),				   \
+				  MODEM_CHAT_MATCH("ERROR", "", NULL));				   \
+	MODEM_CHAT_MATCH_DEFINE(imei_match __maybe_unused,					   \
+				"", "", modem_cellular_chat_on_imei);				   \
+	MODEM_CHAT_MATCH_DEFINE(cgmm_match __maybe_unused,					   \
+				"", "", modem_cellular_chat_on_cgmm);				   \
+	MODEM_CHAT_MATCH_DEFINE(csq_match __maybe_unused,					   \
+				"+CSQ: ", ",", modem_cellular_chat_on_csq);			   \
+	MODEM_CHAT_MATCH_DEFINE(cesq_match __maybe_unused,					   \
+				"+CESQ: ", ",", modem_cellular_chat_on_cesq);			   \
+	MODEM_CHAT_MATCH_DEFINE(qccid_match __maybe_unused,					   \
+				"+QCCID: ", "", modem_cellular_chat_on_iccid);			   \
+	MODEM_CHAT_MATCH_DEFINE(iccid_match __maybe_unused,					   \
+				"+ICCID: ", "", modem_cellular_chat_on_iccid);			   \
+	MODEM_CHAT_MATCH_DEFINE(ccid_match __maybe_unused,					   \
+				"+CCID: ", "", modem_cellular_chat_on_iccid);			   \
+	MODEM_CHAT_MATCH_DEFINE(cimi_match __maybe_unused,					   \
+				"", "", modem_cellular_chat_on_imsi);				   \
+	MODEM_CHAT_MATCH_DEFINE(cgmi_match __maybe_unused,					   \
+				"", "", modem_cellular_chat_on_cgmi);				   \
+	MODEM_CHAT_MATCH_DEFINE(cgmr_match __maybe_unused,					   \
+				"", "", modem_cellular_chat_on_cgmr);				   \
+	MODEM_CHAT_MATCH_DEFINE(connect_match __maybe_unused,					   \
+				"CONNECT", "", NULL);						   \
+	MODEM_CHAT_MATCHES_DEFINE(__maybe_unused abort_matches,					   \
+				  MODEM_CHAT_MATCH("ERROR", "", NULL));				   \
+	MODEM_CHAT_MATCHES_DEFINE(__maybe_unused dial_abort_matches,				   \
+				  MODEM_CHAT_MATCH("ERROR", "", NULL),				   \
+				  MODEM_CHAT_MATCH("BUSY", "", NULL),				   \
+				  MODEM_CHAT_MATCH("NO ANSWER", "", NULL),			   \
+				  MODEM_CHAT_MATCH("NO CARRIER", "", NULL),			   \
+				  MODEM_CHAT_MATCH("NO DIALTONE", "", NULL))
+
 /* Helper to define modem instance */
 #define MODEM_CELLULAR_DEFINE_INSTANCE(inst, power_ms, reset_ms, startup_ms, shutdown_ms, start,   \
 				       set_baudrate_script, init_script, dial_script,              \
